pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-auth"
version = "1.0.0"
description = "FastAPI Authentication API with JWT tokens"
authors = [
    {name = "Lakshay Kapoor", email = "lakshay.netweb@gmail.com"}
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.1",
    "python-multipart>=0.0.6",
    "email-validator>=2.0.0",
]

[project.scripts]
dev = "app.scripts:dev"
start = "app.scripts:start"
prod = "app.scripts:prod"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]